Projectile hit collision is now performed with objLag so that the shooter can aim and hit enemy-players no matter how much lag the shooting player has.
Projectiles is not rendered at client yet though.

+++ trunk/Prototype/src/StandardInterExtraPolate.h
@@ -15,7 +15,7 @@ namespace Prototype
 		{
 			Tickf totalDiff = static_cast<Tickf>(tick2) - static_cast<Tickf>(tick1);
 			Tickf resultDiff = resultTick - static_cast<Tickf>(tick1);
-			float tValue = static_cast<float>(resultDiff) / static_cast<float>(totalDiff);
+			double tValue = static_cast<double>(resultDiff) / static_cast<double>(totalDiff);
 
 			resultData = data1 + (data2 - data1) * tValue;
 		}
